moved MathUtils to utils package
