+++ src/org/opensha2/eq/model/PointSource.java
@@ -5,7 +5,7 @@ import static java.lang.Math.ceil;
 import static org.opensha2.eq.fault.FocalMech.NORMAL;
 import static org.opensha2.eq.fault.FocalMech.REVERSE;
 import static org.opensha2.eq.fault.FocalMech.STRIKE_SLIP;
-import static org.opensha2.internal.MathUtils.hypot;
+import static org.opensha2.util.MathUtils.hypot;
 
 import org.opensha2.data.XySequence;
 import org.opensha2.eq.fault.FocalMech;
