+++ src/org/opensha2/eq/model/PointSourceFixedStrike.java
@@ -8,7 +8,7 @@ import static java.lang.Math.tan;
 
 import static org.opensha2.eq.fault.FocalMech.STRIKE_SLIP;
 import static org.opensha2.geo.GeoTools.TO_RAD;
-import static org.opensha2.internal.MathUtils.hypot;
+import static org.opensha2.util.MathUtils.hypot;
 
 import org.opensha2.data.XySequence;
 import org.opensha2.eq.fault.Faults;
