+++ src/org/opensha2/internal/BinaryCurves.java
@@ -10,6 +10,7 @@ import org.opensha2.geo.GriddedRegion;
 import org.opensha2.geo.Location;
 import org.opensha2.geo.Regions;
 import org.opensha2.gmm.Imt;
+import org.opensha2.util.MathUtils;
 
 import com.google.common.base.Charsets;
 import com.google.common.base.Strings;
